fix for new complier

+++ DMCircularScrollView/DMCircularScrollView/DMCircularScrollView.m
@@ -229,7 +229,7 @@
     }
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)sv withView:(UIView *)view atScale:(float)scale
+- (void)scrollViewDidEndZooming:(UIScrollView *)sv withView:(UIView *)view atScale:(CGFloat)scale
 {
     if (self.scrollViewDelegate && [self.scrollViewDelegate respondsToSelector:@selector(scrollViewDidEndZooming:withView:atScale:)])
     {
